Move configfs functions to new file
The next patches add several new configfs helpers, so move the existing
ones to a new file so it does not clutter api.c.
