+++ libtcmu_common.h
@@ -74,6 +74,7 @@ struct tcmur_handler *tcmu_get_runner_handler(struct tcmu_device *dev);
 /* Helper routines for processing commands */
 int tcmu_get_attribute(struct tcmu_device *dev, const char *name);
 long long tcmu_get_device_size(struct tcmu_device *dev);
+char *tcmu_get_wwn(struct tcmu_device *dev);
 int tcmu_get_cdb_length(uint8_t *cdb);
 uint64_t tcmu_get_lba(uint8_t *cdb);
 uint32_t tcmu_get_xfer_length(uint8_t *cdb);
